Refactored to hexagonal architecture
