+++ ticket-server/pom.xml
@@ -73,6 +73,12 @@
             <artifactId>spring-security-test</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>io.etcd</groupId>
+            <artifactId>jetcd-test</artifactId>
+            <version>0.8.3</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
 </project>
