+++ ticket-server/src/main/java/my/pet/ticket/server/adapter/http/HttpAdapter.java
@@ -1,0 +1,11 @@
+package my.pet.ticket.server.adapter.http;
+
+import my.pet.ticket.server.application.port.entrypoint.EntrypointPort;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HttpAdapter implements EntrypointPort {
+
+
+
+}
