+++ ticket-server/src/main/java/my/pet/ticket/server/adapter/persistence/entity/AbstractEntity.java
@@ -1,0 +1,18 @@
+package my.pet.ticket.server.adapter.persistence.entity;
+
+import jakarta.persistence.Column;
+
+import java.time.ZonedDateTime;
+
+public abstract class AbstractEntity {
+
+    @Column(name = "created_at", nullable = false)
+    private ZonedDateTime createdAt;
+
+    @Column(name = "updated_at", nullable = false)
+    private ZonedDateTime updatedAt;
+
+    @Column(name = "deleted", nullable = false)
+    private Boolean deleted;
+
+}
