+++ ticket-server/src/main/java/my/pet/ticket/server/adapter/persistence/entity/ClientEntity.java
@@ -1,0 +1,46 @@
+package my.pet.ticket.server.adapter.persistence.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Data;
+
+import java.util.Objects;
+
+@Data
+@Entity
+@Table(schema = "pet_project", name = "clients")
+public class ClientEntity extends AbstractEntity {
+
+    @Id
+    @Column(name = "client_id")
+    private Long clientId;
+
+    @Column(name = "first_name")
+    private String firstName;
+
+    @Column(name = "last_name")
+    private String lastName;
+
+    @Column(name = "sur_name")
+    private String surName;
+
+    @Column(name = "full_name")
+    private String fullName;
+
+    @Column(name = "email")
+    private String email;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof ClientEntity that)) return false;
+        return Objects.equals(clientId, that.clientId) && Objects.equals(firstName, that.firstName) && Objects.equals(lastName, that.lastName) && Objects.equals(surName, that.surName) && Objects.equals(email, that.email);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(clientId, firstName, lastName, surName, email);
+    }
+}
