+++ ticket-server/src/main/java/my/pet/ticket/server/adapter/persistence/repository/ClientRepository.java
@@ -1,0 +1,9 @@
+package my.pet.ticket.server.adapter.persistence.repository;
+
+import my.pet.ticket.server.adapter.persistence.entity.ClientEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ClientRepository extends CrudRepository<ClientEntity, Long> {
+}
