+++ ticket-server/src/main/java/my/pet/ticket/server/adapter/persistence/repository/PhoneNumberRepository.java
@@ -1,0 +1,10 @@
+package my.pet.ticket.server.adapter.persistence.repository;
+
+import my.pet.ticket.server.adapter.persistence.entity.PhoneNumberEntity;
+import my.pet.ticket.server.adapter.persistence.entity.PhoneNumberIdEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PhoneNumberRepository extends CrudRepository<PhoneNumberEntity, PhoneNumberIdEntity> {
+}
