+++ ticket-server/src/main/java/my/pet/ticket/server/adapter/persistence/repository/RolesRepository.java
@@ -1,0 +1,9 @@
+package my.pet.ticket.server.adapter.persistence.repository;
+
+import my.pet.ticket.server.adapter.persistence.entity.RoleEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RolesRepository extends CrudRepository<RoleEntity, Long> {
+}
