+++ ticket-server/src/main/java/my/pet/ticket/server/adapter/persistence/repository/TicketRepository.java
@@ -1,0 +1,10 @@
+package my.pet.ticket.server.adapter.persistence.repository;
+
+import my.pet.ticket.server.adapter.persistence.entity.TicketEntity;
+import my.pet.ticket.server.adapter.persistence.entity.TicketIdEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TicketRepository extends CrudRepository<TicketEntity, TicketIdEntity> {
+}
