+++ ticket-server/src/main/java/my/pet/ticket/server/adapter/persistence/repository/TicketStatusRepository.java
@@ -1,0 +1,9 @@
+package my.pet.ticket.server.adapter.persistence.repository;
+
+import my.pet.ticket.server.adapter.persistence.entity.TicketStatusEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TicketStatusRepository extends CrudRepository<TicketStatusEntity, Long> {
+}
