+++ ticket-server/src/main/java/my/pet/ticket/server/adapter/persistence/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package my.pet.ticket.server.adapter.persistence.repository;
+
+import my.pet.ticket.server.adapter.persistence.entity.UserEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<UserEntity, Long> {
+}
