+++ ticket-server/src/main/java/my/pet/ticket/server/application/domain/model/Client.java
@@ -1,0 +1,16 @@
+package my.pet.ticket.server.application.domain.model;
+
+import lombok.Data;
+
+@Data
+public class Client {
+
+    private Long clientId;
+
+    private String fullName;
+
+    private String email;
+
+    private String phoneNumber;
+
+}
