+++ ticket-server/src/main/java/my/pet/ticket/server/application/domain/model/Role.java
@@ -1,0 +1,12 @@
+package my.pet.ticket.server.application.domain.model;
+
+import lombok.Data;
+
+@Data
+public class Role {
+
+    private Long roleId;
+
+    private String name;
+
+}
