+++ ticket-server/src/main/java/my/pet/ticket/server/application/port/entrypoint/EntrypointPort.java
@@ -1,0 +1,7 @@
+package my.pet.ticket.server.application.port.entrypoint;
+
+public interface EntrypointPort {
+
+
+
+}
