+++ ticket-server/src/main/java/my/pet/ticket/server/common/logging/Detail.java
@@ -1,0 +1,9 @@
+package my.pet.ticket.server.common.logging;
+
+public interface Detail {
+
+    String getErrorCode();
+
+    String getDescription();
+
+}
