+++ ticket-server/src/main/java/my/pet/ticket/server/common/logging/Event.java
@@ -1,0 +1,9 @@
+package my.pet.ticket.server.common.logging;
+
+import com.fasterxml.jackson.databind.node.ObjectNode;
+
+public interface Event {
+
+    ObjectNode toDetails();
+
+}
