+++ ticket-server/src/main/java/my/pet/ticket/server/common/logging/LoggingException.java
@@ -1,0 +1,9 @@
+package my.pet.ticket.server.common.logging;
+
+public class LoggingException extends RuntimeException {
+
+    public LoggingException(Throwable e) {
+        super(e);
+    }
+
+}
