--- ticket-server/src/main/java/my/pet/ticket/server/dao/AbstractDao.java
@@ -1,18 +1,0 @@
-package my.pet.ticket.server.dao;
-
-import jakarta.persistence.Column;
-
-import java.time.ZonedDateTime;
-
-public abstract class AbstractDao {
-
-    @Column(name = "created_at", nullable = false)
-    private ZonedDateTime createdAt;
-
-    @Column(name = "updated_at", nullable = false)
-    private ZonedDateTime updatedAt;
-
-    @Column(name = "deleted", nullable = false)
-    private Boolean deleted;
-
-}
