--- ticket-server/src/main/java/my/pet/ticket/server/dao/TicketIdDao.java
@@ -1,17 +1,0 @@
-package my.pet.ticket.server.dao;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.Data;
-
-@Data
-@Embeddable
-public class TicketIdDao {
-
-    @Column(name = "ticket_id")
-    private Long ticketId;
-
-    @Column(name = "client_id")
-    private Long clientId;
-
-}
