--- ticket-server/src/main/java/my/pet/ticket/server/logging/DefaultDetail.java
@@ -1,21 +1,0 @@
-package my.pet.ticket.server.logging;
-
-import lombok.Data;
-
-@Data
-public class DefaultDetail implements Detail {
-
-    private final String errorCode;
-
-    private final String description;
-
-    public DefaultDetail(String errorCode) {
-        this(errorCode, null);
-    }
-
-    public DefaultDetail(String errorCode, String description) {
-        this.errorCode = errorCode;
-        this.description = description;
-    }
-
-}
