--- ticket-server/src/main/java/my/pet/ticket/server/logging/Detail.java
@@ -1,9 +1,0 @@
-package my.pet.ticket.server.logging;
-
-public interface Detail {
-
-    String getErrorCode();
-
-    String getDescription();
-
-}
