--- ticket-server/src/main/java/my/pet/ticket/server/logging/Event.java
@@ -1,9 +1,0 @@
-package my.pet.ticket.server.logging;
-
-import com.fasterxml.jackson.databind.node.ObjectNode;
-
-public interface Event {
-
-    ObjectNode toDetails();
-
-}
