--- ticket-server/src/main/java/my/pet/ticket/server/logging/Events.java
@@ -1,23 +1,0 @@
-package my.pet.ticket.server.logging;
-
-import lombok.Getter;
-
-@Getter
-public enum Events {
-
-    BP_START("Business process start"),
-
-    BP_IN_PROGRESS("Business process in progress"),
-
-    BP_END("Business process ended");
-
-    private final String name;
-
-    Events(String name) {
-        this.name = name;
-    }
-
-    public Event getEvent() {
-        return new DefaultEvent(name);
-    }
-}
