--- ticket-server/src/main/java/my/pet/ticket/server/logging/LoggingException.java
@@ -1,9 +1,0 @@
-package my.pet.ticket.server.logging;
-
-public class LoggingException extends RuntimeException {
-
-    public LoggingException(Throwable e) {
-        super(e);
-    }
-
-}
