--- ticket-server/src/main/java/my/pet/ticket/server/repository/TicketRepository.java
@@ -1,10 +1,0 @@
-package my.pet.ticket.server.repository;
-
-import my.pet.ticket.server.dao.TicketDao;
-import my.pet.ticket.server.dao.TicketIdDao;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TicketRepository extends CrudRepository<TicketDao, TicketIdDao> {
-}
