--- ticket-server/src/main/java/my/pet/ticket/server/repository/TicketStatusRepository.java
@@ -1,9 +1,0 @@
-package my.pet.ticket.server.repository;
-
-import my.pet.ticket.server.dao.TicketStatusDao;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TicketStatusRepository extends CrudRepository<TicketStatusDao, Long> {
-}
