+++ ticket-server/src/test/java/my/pet/ticket/server/application/port/configuration/ConfigurationPortTest.java
@@ -1,0 +1,21 @@
+package my.pet.ticket.server.application.port.configuration;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@SpringBootTest
+class ConfigurationPortTest {
+
+    @Autowired
+    private ConfigurationPort configurationPort;
+
+    @Test
+    void getFirst() {
+        TestValue actual = configurationPort.getFirst(TestValue.class, "test", null);
+        Assertions.assertEquals(new TestValue("test"), actual);
+    }
+}
