+++ ticket-server/src/test/java/my/pet/ticket/server/application/port/configuration/TestValue.java
@@ -1,0 +1,16 @@
+package my.pet.ticket.server.application.port.configuration;
+
+import com.fasterxml.jackson.annotation.JsonAlias;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class TestValue {
+
+    @JsonAlias("value")
+    private String value;
+
+}
